--- flink/src/test/java/io/delta/flink/DeltaTestUtils.java
@@ -1,17 +1,0 @@
-package io.delta.flink;
-
-import org.apache.hadoop.conf.Configuration;
-
-public class DeltaTestUtils {
-
-    ///////////////////////////////////////////////////////////////////////////
-    // hadoop conf test utils
-    ///////////////////////////////////////////////////////////////////////////
-
-    public static Configuration getHadoopConf() {
-        Configuration conf = new Configuration();
-        conf.set("parquet.compression", "SNAPPY");
-        return conf;
-    }
-
-}
