+++ flink/src/test/java/io/delta/flink/sink/DeltaSinkStreamingExecutionITCase.java
@@ -32,7 +32,8 @@ import java.util.concurrent.CountDownLatch;
 import java.util.stream.LongStream;
 
 import io.delta.flink.sink.utils.DeltaSinkTestUtils;
-import io.delta.flink.sink.utils.TestParquetReader;
+import io.delta.flink.utils.DeltaTestUtils;
+import io.delta.flink.utils.TestParquetReader;
 import org.apache.flink.api.common.RuntimeExecutionMode;
 import org.apache.flink.api.common.restartstrategy.RestartStrategies;
 import org.apache.flink.api.common.state.CheckpointListener;
@@ -98,9 +99,9 @@ public class DeltaSinkStreamingExecutionITCase extends StreamingExecutionFileSin
         try {
             deltaTablePath = TEMPORARY_FOLDER.newFolder().getAbsolutePath();
             if (isPartitioned) {
-                DeltaSinkTestUtils.initTestForPartitionedTable(deltaTablePath);
+                DeltaTestUtils.initTestForPartitionedTable(deltaTablePath);
             } else {
-                DeltaSinkTestUtils.initTestForNonPartitionedTable(deltaTablePath);
+                DeltaTestUtils.initTestForNonPartitionedTable(deltaTablePath);
             }
         } catch (IOException e) {
             throw new RuntimeException("Weren't able to setup the test dependencies", e);
@@ -120,7 +121,7 @@ public class DeltaSinkStreamingExecutionITCase extends StreamingExecutionFileSin
 
     public void runDeltaSinkTest() throws Exception {
         // GIVEN
-        DeltaLog deltaLog = DeltaLog.forTable(DeltaSinkTestUtils.getHadoopConf(), deltaTablePath);
+        DeltaLog deltaLog = DeltaLog.forTable(DeltaTestUtils.getHadoopConf(), deltaTablePath);
         List<AddFile> initialDeltaFiles = deltaLog.snapshot().getAllFiles();
         long initialVersion = deltaLog.snapshot().getVersion();
         int initialTableRecordsCount = TestParquetReader.readAndValidateAllTableRecords(deltaLog);
