+++ flink/src/test/java/io/delta/flink/sink/internal/writer/DeltaWriterBucketTest.java
@@ -30,7 +30,7 @@ import java.util.stream.Stream;
 import io.delta.flink.sink.internal.committables.DeltaCommittable;
 import io.delta.flink.sink.internal.committer.DeltaCommitter;
 import io.delta.flink.sink.utils.DeltaSinkTestUtils;
-import io.delta.flink.sink.utils.TestParquetReader;
+import io.delta.flink.utils.TestParquetReader;
 import org.apache.flink.core.fs.Path;
 import org.apache.flink.metrics.Counter;
 import org.apache.flink.metrics.MetricGroup;
